Protocol update to make it more readable
Instead of sending a [success-flags, result-list] array, the response
now consists of a single [success-flag, result]-tuple for every request.
This, of course, has the drawback of adding a few more bytes to memory
usage on both sides (more arrays), but the overhead is really negligable
compared to the readability gain of the protocol.

Also switched to score.js.exc2json for transporting/formatting
exceptions.